Fixed reference comparison of Input objects
The 'Input' class is designed like a value object, the „Equals“ method
is overriden, but in some cases it was mistakenly not used for
comparison. For instance, in the method „XOr<T>“ line „if
(ff.FailedInput != i)“ used reference comparison, where it seems must be
value comparison.

So I added assertions to the  tests, to test the eq. operators too, and
overrided the  operators to make tests succeed.