Optimize the Parse.Regex method
Create a new Regex based on the Regex parameter. Wrap the original pattern in non-capturing
parentheses preceded by a `^', while preserving the original Regex's options. This way, the
regex engine won't attempt to match the pattern at any position other than the beginning of
the input. This is done to match the way that Sprache parses input. Previously, a check was
made in the delegate returned by the Parse.Regex method in order to verify that the match's
Index was zero. This check is no longer necessary.

This is essentially what happens to the Regex parameter:

```c#
regex = new Regex(string.Format("^(?:{0})", regex), regex.Options);
```